+++ extensions/writersbrew/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Writersbrew Changelog
 
+## [New Commands] - 2023-01-24
+
+- Added Explain code
+- Added Capture text
+
 ## [New Commands] - 2023-01-17
 
 - Added Translate to dutch